feat: 메인 페이지 UI 개선 및 memberInfo 권한 처리 수정
- 메인 홈페이지 이미지 교체 (로컬 이미지 carousel 적용)
- 최신글/인기글/가이드 게시판 데이터 fetch 로직 추가 및 콘솔 확인
- 다이어리 페이지: 자기 다이어리일 경우 친구추가 버튼 노출 안 되도록 조건 수정
- 회원정보 수정(MemberEdit): memberInfo.id와 params.id 비교 시 타입 일치 처리
- 관리자 전용 페이지 접근 제어: role 값 정규화("ROLE_ADMIN" → "ADMIN") 적용
- LoginProvider: fetchMemberInfo에서 role 정규화 및 memberInfo 구조 안정화
- Axios 인터셉터 및 토큰 만료 처리 로직 점검/보완

+++ backend/src/main/java/com/backend/service/diary/DiaryService.java
@@ -47,7 +47,9 @@ public class DiaryService {
         return diaryMapper.selectByMemberId(memberId);
     }
 
-    public Diary getDiaryById(Long id) {
+    public Diary getDiaryById(Integer id) {
         return diaryMapper.selectById(id);
     }
+
+
 }
